updated Readme to have non-collapsing whitespace, and breaks
It is required to add special whitespaces in front of text to get it to indent.  Tab doesn't do it.    
It is required to add 4 spaces at the end of a line to get a line-break effect.  Carriage-return doesn't do it.

I cleaned up the part of the text in the "config file" area.